Change path for register and move registration method to security controller. Working on new user creation and last verification.
