+++ src/main/java/com/fradou/nutrition/mvc/dao/generic/GenericDAO.java
@@ -20,7 +20,7 @@ public interface GenericDAO<T> {
 	public int count();
 	
 	public T findUniqueBy(String field, String value);
-	
+		
 	public List<T> find(Integer offset, Integer entries, String orderBy, String orderType);
 	
 }
