+++ src/main/java/com/fradou/nutrition/mvc/entity/CustomUser.java
@@ -209,6 +209,8 @@ public class CustomUser implements UserDetails {
 		this.roles = roles;
 	}
 	
+	
+	
 	public String toString() {
 		return "CustomerUser : [id=" + id + ", username=" + username + ", height="+height+", age=" + age;
 	}
